Move `validatePackageName` to `jsTyping.ts`
